✨ feat: 모바일 대응 랜딩 페이지 UI 추가 및 feed 페이지 폴더 구조 변경
