+++ nova-app/src/app/(protected)/feed/ui/FeedArticle.tsx
@@ -1,15 +1,13 @@
 'use client';
 
 import { ArticleCard } from '@/shared/ui';
-import {
-  FeedArticleSkeleton as ArticleCardSkeleton,
-  FeedArticleError,
-  EmptyFeed,
-} from '@/app/(protected)/feed/ui';
-import { useInfiniteSavedArticles } from '@/features/saved/hooks/useInfiniteSavedArticles';
+import ArticleCardSkeleton from './FeedArticleSkeleton';
+import FeedArticleError from './FeedArticleError';
+import EmptyFeed from './EmptyFeed';
+import { useInfiniteFeedArticles } from '@/features/feed/hooks/useInfiniteFeedArticles';
 import { useInfiniteScroll } from '@/shared/hooks';
 
-const SavedArticleList = () => {
+const FeedArticle = () => {
   const {
     articles,
     isLoading,
@@ -18,7 +16,7 @@ const SavedArticleList = () => {
     hasNextPage,
     fetchNextPage,
     isFetchingNextPage,
-  } = useInfiniteSavedArticles();
+  } = useInfiniteFeedArticles();
 
   const { targetRef } = useInfiniteScroll({
     hasNextPage: hasNextPage ?? false,
@@ -52,7 +50,7 @@ const SavedArticleList = () => {
 
   return (
     <>
-      <section id='saved-articles-export-root' className='space-y-4'>
+      <section className='space-y-4'>
         {articles.map((article) => (
           <ArticleCard key={article.id} articleData={article} />
         ))}
@@ -63,11 +61,10 @@ const SavedArticleList = () => {
             ))}
           </div>
         )}
-
         {!hasNextPage && !isFetchingNextPage && articles.length > 0 && (
-          <div className='flex justify-center items-center py-2'>
+          <div className='flex justify-center items-center py-8'>
             <p className='typo-body-base text-additive'>
-              더 이상 저장한 콘텐츠가 없습니다.
+              더 이상 피드 콘텐츠가 없습니다.
             </p>
           </div>
         )}
@@ -77,4 +74,4 @@ const SavedArticleList = () => {
   );
 };
 
-export default SavedArticleList;
+export default FeedArticle;
