+++ nova-app/src/app/(protected)/feed/ui/FeedFilterSidebar.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import { PeriodFilter, SortFilter, TypeFilter, KeywordFilter } from '@/features/feed/ui';
+
+const FeedFilterSidebar = () => {
+  return (
+    <aside className='w-80 space-y-4 max-lg:hidden'>
+      <SortFilter />
+      <PeriodFilter />
+      <TypeFilter />
+      <KeywordFilter />
+    </aside>
+  );
+};
+
+export default FeedFilterSidebar;
