+++ nova-app/src/features/feed/hooks/index.ts
@@ -3,3 +3,4 @@ export { useFeedArticles } from './useFeedArticles';
 export { useHideFeed } from './useHideFeed';
 export { useInfiniteFeedArticles } from './useInfiniteFeedArticles';
 
+
